+++ packages/firebase-auth/src/index.ts
@@ -1,6 +1,7 @@
 import type { KeyStorer, FirebaseIdToken } from 'firebase-auth-cloudflare-workers'
 import { Auth, WorkersKVStoreSingle } from 'firebase-auth-cloudflare-workers'
 import type { Context, MiddlewareHandler } from 'hono'
+import { getCookie } from 'hono/cookie'
 import { HTTPException } from 'hono/http-exception'
 
 export type VerifyFirebaseAuthEnv = {
@@ -21,10 +22,10 @@ export interface VerifyFirebaseAuthConfig {
 const defaultKVStoreJWKCacheKey = 'verify-firebase-auth-cached-public-key'
 const defaultKeyStoreInitializer = (c: Context<{ Bindings: VerifyFirebaseAuthEnv }>): KeyStorer => {
   if (c.env.PUBLIC_JWK_CACHE_KV === undefined) {
-    const res = new Response('Not Implemented', {
-      status: 501,
+    const status = 501
+    throw new HTTPException(status, {
+      res: new Response('Not Implemented', { status }),
     })
-    throw new HTTPException(501, { res })
   }
   return WorkersKVStoreSingle.getOrInitialize(
     c.env.PUBLIC_JWK_CACHE_KEY ?? defaultKVStoreJWKCacheKey,
@@ -35,29 +36,33 @@ const defaultKeyStoreInitializer = (c: Context<{ Bindings: VerifyFirebaseAuthEnv
 export const verifyFirebaseAuth = (userConfig: VerifyFirebaseAuthConfig): MiddlewareHandler => {
   const config = {
     projectId: userConfig.projectId,
-    AuthorizationHeaderKey: userConfig.authorizationHeaderKey ?? 'Authorization',
-    KeyStore: userConfig.keyStore,
+    authorizationHeaderKey: userConfig.authorizationHeaderKey ?? 'Authorization',
+    keyStore: userConfig.keyStore,
     keyStoreInitializer: userConfig.keyStoreInitializer ?? defaultKeyStoreInitializer,
     disableErrorLog: userConfig.disableErrorLog,
     firebaseEmulatorHost: userConfig.firebaseEmulatorHost,
-  }
+  } satisfies VerifyFirebaseAuthConfig
+
+  // TODO(codehex): will be supported
+  const checkRevoked = false
 
   return async (c, next) => {
-    const authorization = c.req.headers.get(config.AuthorizationHeaderKey)
+    const authorization = c.req.raw.headers.get(config.authorizationHeaderKey)
     if (authorization === null) {
-      const res = new Response('Bad Request', {
-        status: 400,
+      const status = 400
+      throw new HTTPException(status, {
+        res: new Response('Bad Request', { status }),
+        message: 'authorization header is empty',
       })
-      throw new HTTPException(400, { res })
     }
     const jwt = authorization.replace(/Bearer\s+/i, '')
     const auth = Auth.getOrInitialize(
       config.projectId,
-      config.KeyStore ?? config.keyStoreInitializer(c)
+      config.keyStore ?? config.keyStoreInitializer(c)
     )
 
     try {
-      const idToken = await auth.verifyIdToken(jwt, {
+      const idToken = await auth.verifyIdToken(jwt, checkRevoked, {
         FIREBASE_AUTH_EMULATOR_HOST:
           config.firebaseEmulatorHost ?? c.env.FIREBASE_AUTH_EMULATOR_HOST,
       })
@@ -70,10 +75,12 @@ export const verifyFirebaseAuth = (userConfig: VerifyFirebaseAuthConfig): Middle
         })
       }
 
-      const res = new Response('Unauthorized', {
-        status: 401,
+      const status = 401
+      throw new HTTPException(status, {
+        res: new Response('Unauthorized', { status }),
+        message: `failed to verify the requested firebase token: ${String(err)}`,
+        cause: err,
       })
-      throw new HTTPException(401, { res })
     }
     await next()
   }
@@ -89,4 +96,61 @@ export const getFirebaseToken = (c: Context): FirebaseIdToken | null => {
     return null
   }
   return idToken
+}
+
+export interface VerifySessionCookieFirebaseAuthConfig {
+  projectId: string
+  cookieName?: string
+  keyStore?: KeyStorer
+  keyStoreInitializer?: (c: Context) => KeyStorer
+  firebaseEmulatorHost?: string
+  redirects: {
+    signIn: string
+  }
+}
+
+export const verifySessionCookieFirebaseAuth = (
+  userConfig: VerifySessionCookieFirebaseAuthConfig
+): MiddlewareHandler => {
+  const config = {
+    projectId: userConfig.projectId,
+    cookieName: userConfig.cookieName ?? 'session',
+    keyStore: userConfig.keyStore,
+    keyStoreInitializer: userConfig.keyStoreInitializer ?? defaultKeyStoreInitializer,
+    firebaseEmulatorHost: userConfig.firebaseEmulatorHost,
+    redirects: userConfig.redirects,
+  } satisfies VerifySessionCookieFirebaseAuthConfig
+
+  // TODO(codehex): will be supported
+  const checkRevoked = false
+
+  return async (c, next) => {
+    const auth = Auth.getOrInitialize(
+      config.projectId,
+      config.keyStore ?? config.keyStoreInitializer(c)
+    )
+    const session = getCookie(c, config.cookieName)
+    if (session === undefined) {
+      const status = 302
+      const res = c.redirect(config.redirects.signIn, status)
+      throw new HTTPException(status, { res, message: 'session is empty' })
+    }
+
+    try {
+      const idToken = await auth.verifySessionCookie(session, checkRevoked, {
+        FIREBASE_AUTH_EMULATOR_HOST:
+          config.firebaseEmulatorHost ?? c.env.FIREBASE_AUTH_EMULATOR_HOST,
+      })
+      setFirebaseToken(c, idToken)
+    } catch (err) {
+      const status = 302
+      const res = c.redirect(config.redirects.signIn, status)
+      throw new HTTPException(status, {
+        res,
+        message: `failed to verify the requested firebase token: ${String(err)}`,
+        cause: err,
+      })
+    }
+    await next()
+  }
 }