+++ packages/react-renderer/src/react-renderer.ts
@@ -1,12 +1,13 @@
 import type { Context } from 'hono'
 import type { Env, MiddlewareHandler } from 'hono/types'
 import React from 'react'
-import { renderToString } from 'react-dom/server'
+import { renderToString, type RenderToReadableStreamOptions } from 'react-dom/server'
 import type { Props } from '.'
 
 type RendererOptions = {
   docType?: boolean | string
   stream?: boolean | Record<string, string>
+  readableStreamOptions?: RenderToReadableStreamOptions
 }
 
 type BaseProps = {
@@ -14,29 +15,37 @@ type BaseProps = {
   children: React.ReactElement
 }
 
+type ComponentProps = Props &
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  BaseProps & { Layout: React.FC<Record<string, any> & { children: React.ReactElement }> }
+
 const RequestContext = React.createContext<Context | null>(null)
 
 const createRenderer =
-  (c: Context, component?: React.FC<Props & BaseProps>, options?: RendererOptions) =>
+  (
+    c: Context,
+    Layout: React.FC<{ children: React.ReactElement }>,
+    component?: React.FC<ComponentProps>,
+    options?: RendererOptions
+  ) =>
   async (children: React.ReactElement, props?: Props) => {
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    const node = component ? component({ children, c, ...props }) : children
+    const node = component ? component({ children, Layout, c, ...props }) : children
 
     if (options?.stream) {
       const { renderToReadableStream } = await import('react-dom/server')
       const stream = await renderToReadableStream(
-        React.createElement(RequestContext.Provider, { value: c }, node)
+        React.createElement(RequestContext.Provider, { value: c }, node),
+        options.readableStreamOptions
       )
-      return c.body(stream, {
-        headers:
-          options.stream === true
-            ? {
-                'Transfer-Encoding': 'chunked',
-                'Content-Type': 'text/html; charset=UTF-8',
-              }
-            : options.stream,
-      })
+      if (options.stream === true) {
+        c.header('Transfer-Encoding', 'chunked')
+        c.header('Content-Type', 'text/html; charset=UTF-8')
+      } else {
+        for (const [key, value] of Object.entries(options.stream)) {
+          c.header(key, value)
+        }
+      }
+      return c.body(stream)
     } else {
       const docType =
         typeof options?.docType === 'string'
@@ -51,11 +60,20 @@ const createRenderer =
   }
 
 export const reactRenderer = (
-  component?: React.FC<Props & BaseProps>,
+  component?: React.FC<ComponentProps>,
   options?: RendererOptions
 ): MiddlewareHandler =>
   function reactRenderer(c, next) {
-    c.setRenderer(createRenderer(c, component, options))
+    const Layout = (c.getLayout() ?? React.Fragment) as React.FC<{
+      children: React.ReactElement
+    }>
+    if (component) {
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      c.setLayout((props: any) => {
+        return component({ ...props, Layout, c }, c)
+      })
+    }
+    c.setRenderer(createRenderer(c, Layout, component, options))
     return next()
   }